requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keras==3.0.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.3.1
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.3
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
scikit-learn==1.7.2
scipy==1.12.0
sentry-sdk==2.42.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tensorboard==2.16.0
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
